chore: 파일 그룹 구조 변경
